move `tools` into `utils`
