+++ frb_codegen/src/error.rs
@@ -1,6 +1,6 @@
 use thiserror::Error;
 
-use crate::tools::PackageManager;
+use crate::utils::tools::PackageManager;
 
 pub type Result<T = ()> = std::result::Result<T, Error>;
