+++ frb_codegen/src/lib.rs
@@ -16,10 +16,8 @@ use crate::commands::BindgenRustToDartArg;
 use crate::others::*;
 use crate::target::Acc;
 use crate::target::Target;
-use crate::utils::*;
 
 mod config;
-mod tools;
 
 pub use crate::commands::ensure_tools_available;
 pub use crate::config::parse as config_parse;
