+++ frb_codegen/src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod consts;
 pub(crate) mod misc;
+pub(crate) mod tools;
